fix dimen

+++ imagepickerview/src/main/java/ru/vvdev/imagepickerview/ImageAttr.kt
@@ -2,8 +2,8 @@ package ru.vvdev.imagepickerview
 
 class ImageAttr(
         val cornerRadius: Float,
-        val viewHeight: Float,
-        val viewWight: Float,
+        val viewHeight: Int,
+        val viewWight: Int,
         var backClose: Int,
         var tintClose: Int,
         val addAttr: AddAttr
